@@ -13,135 +24,136 @@
     "@react-navigation/bottom-tabs": "^6.0.9",
     "@react-navigation/native": "^6.0.6",
     "@react-navigation/stack": "^6.2.2",
-    "@walletconnect/react-native-dapp": "1.6.1",
+    "@tradle/react-native-http": "^2.0.1",
+    "assert": "^2.0.0",
     "base-64": "1.0.0",
     "bip39": "^3.0.4",
+    "browserify-zlib": "^0.1.4",
     "buffer": "6.0.3",
+    "console-browserify": "^1.2.0",
+    "constants-browserify": "^1.0.0",
+    "dns.js": "^1.0.1",
+    "domain-browser": "^1.2.0",
     "ethereumjs-util": "^7.1.3",
     "ethereumjs-wallet": "^1.0.2",
-    "expo": "^44.0.0",
-    "expo-splash-screen": "~0.14.1",
-    "expo-status-bar": "~1.2.0",
+    "ethers": "^5.6.9",
+    "https-browserify": "^0.0.1",
+    "lodash": "^4.17.21",
     "node-libs-browser": "2.2.1",
-    "path-browserify": "0.0.0",
+    "path-browserify": "^0.0.0",
+    "punycode": "^1.4.1",
+    "querystring-es3": "^0.2.1",
     "randombytes": "^2.1.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-native": "0.64.3",
-    "react-native-crypto": "2.2.0",
+    "react": "18.0.0",
+    "react-native": "0.69.1",
+    "react-native-crypto": "^2.2.0",
     "react-native-dotenv": "2.4.3",
     "react-native-gesture-handler": "^2.5.0",
-    "react-native-get-random-values": "1.5.0",
-    "react-native-localhost": "1.0.0",
-    "react-native-safe-area-context": "3.3.2",
+    "react-native-get-random-values": "^1.8.0",
+    "react-native-level-fs": "^3.0.1",
+    "react-native-os": "^1.2.6",
+    "react-native-randombytes": "^3.6.1",
+    "react-native-safe-area-context": "^4.3.1",
     "react-native-screens": "~3.10.1",
-    "react-native-stream": "0.1.9",
-    "react-native-svg": "9.6.4",
-    "react-native-web": "0.17.1",
+    "react-native-stream": "^0.1.9",
+    "react-native-svg": "12.0.3",
+    "react-native-tcp": "^3.3.2",
+    "react-native-udp": "^2.7.0",
+    "readable-stream": "^1.0.33",
+    "rn-nodeify": "^10.3.0",
+    "stream-browserify": "^1.0.0",
+    "string_decoder": "^0.10.31",
+    "timers-browserify": "^1.4.2",
+    "tty-browserify": "^0.0.0",
     "twrnc": "^3.3.3",
+    "url": "^0.10.3",
+    "util": "^0.10.4",
+    "vm-browserify": "^0.0.4",
     "web3": "1.3.1"
   },
   "devDependencies": {
-    "@babel/core": "7.15.5",
-    "@babel/plugin-proposal-private-property-in-object": "7.15.4",
-    "@babel/preset-env": "7.15.6",
-    "@babel/preset-typescript": "7.15.0",
-    "@nomiclabs/hardhat-ethers": "^2.0.1",
-    "@nomiclabs/hardhat-waffle": "^2.0.1",
-    "@openzeppelin/contracts": "^4.3.0",
-    "@types/jest": "^26.0.20",
-    "@types/node": "14.14.22",
-    "@types/react": "17.0.1",
-    "@types/react-native": "~0.64.12",
-    "@typescript-eslint/eslint-plugin": "^4.0.1",
-    "@typescript-eslint/parser": "^4.0.1",
-    "app-root-path": "3.0.0",
-    "babel-jest": "27.2.3",
-    "chai": "^4.2.0",
-    "chokidar": "3.5.1",
-    "commitizen": "4.2.3",
-    "cz-conventional-changelog": "^3.2.0",
-    "dotenv": "8.2.0",
-    "enzyme": "3.11.0",
-    "enzyme-adapter-react-16": "1.15.6",
-    "eslint": "^7.8.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-functional": "^3.0.2",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "7.22.0",
-    "eslint-plugin-react-native": "3.10.0",
-    "ethereum-waffle": "^3.2.1",
-    "gh-pages": "^3.2.3",
-    "hardhat": "2.0.6",
-    "husky": "4.3.8",
-    "jest": "26.6.3",
-    "lint-staged": "10.5.3",
-    "platform-detect": "3.0.1",
-    "prettier": "^2.5.1",
-    "react-test-renderer": "17.0.1",
-    "ts-node": "9.1.1",
-    "typescript": "~4.3.5"
+    "@babel/core": "^7.12.9",
+    "@babel/runtime": "^7.12.5",
+    "@react-native-community/eslint-config": "^2.0.0",
+    "@tsconfig/react-native": "^2.0.0",
+    "@types/jest": "^26.0.23",
+    "@types/lodash": "^4.14.182",
+    "@types/react-native": "^0.69.1",
+    "@types/react-test-renderer": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
+    "babel-jest": "^26.6.3",
+    "eslint": "^7.32.0",
+    "jest": "^26.6.3",
+    "metro-react-native-babel-preset": "^0.70.3",
+    "react-test-renderer": "18.0.0",
+    "typescript": "^4.4.4"
   },
-  "scripts": {
-    "start": "react-native start",
-    "android": "react-native run-android",
-    "ios": "react-native run-ios",
-    "web": "node_modules/.bin/ts-node scripts/web",
-    "eject": "expo eject",
-    "audit": "npx yarn-audit-fix",
-    "postinstall": "node_modules/.bin/ts-node scripts/postinstall",
-    "test": "npx hardhat test && jest --passWithNoTests",
-    "web:deploy": "expo build:web && gh-pages -d web-build"
-  },
-  "version": "1.0.0",
-  "private": true,
-  "name": "sled",
-  "homepage": "https://cawfree.github.io/create-react-native-dapp",
-  "license": "MIT",
-  "contributors": [
-    {
-      "name": "@cawfree",
-      "url": "https://github.com/cawfree"
-    }
-  ],
-  "keywords": [
-    "react",
-    "react-native",
-    "blockchain",
-    "dapp",
-    "ethereum",
-    "web3",
-    "starter",
-    "react-native-web"
-  ],
-  "react-native": {
-    "stream": "react-native-stream",
-    "crypto": "react-native-crypto",
-    "path": "path-browserify",
-    "process": "node-libs-browser/mock/process"
+  "resolutions": {
+    "@types/react": "^18"
   },
   "jest": {
     "preset": "react-native",
-    "testMatch": [
-      "**/__tests__/src/**/*.[jt]s?(x)"
-    ],
-    "transformIgnorePatterns": [
-      "node_modules/(?!@ngrx|(?!deck.gl)|ng-dynamic)"
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "jsx",
+      "json",
+      "node"
     ]
   },
-  "config": {
-    "commitizen": {
-      "path": "./node_modules/cz-conventional-changelog"
-    }
+  "react-native": {
+    "console": "console-browserify",
+    "path": "path-browserify",
+    "fs": "react-native-level-fs",
+    "dgram": "react-native-udp",
+    "zlib": "browserify-zlib",
+    "constants": "constants-browserify",
+    "crypto": "react-native-crypto",
+    "dns": "dns.js",
+    "net": "react-native-tcp",
+    "domain": "domain-browser",
+    "http": "@tradle/react-native-http",
+    "https": "https-browserify",
+    "os": "react-native-os",
+    "querystring": "querystring-es3",
+    "_stream_transform": "readable-stream/transform",
+    "_stream_readable": "readable-stream/readable",
+    "_stream_writable": "readable-stream/writable",
+    "_stream_duplex": "readable-stream/duplex",
+    "_stream_passthrough": "readable-stream/passthrough",
+    "stream": "stream-browserify",
+    "timers": "timers-browserify",
+    "tty": "tty-browserify",
+    "vm": "vm-browserify",
+    "tls": false,
+    "process": "node-libs-browser/mock/process"
   },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged && tsc"
-    }
-  },
-  "lint-staged": {
-    "*.{ts,tsx,js,jsx}": "eslint --fix --ext '.ts,.tsx,.js,.jsx' -c .eslintrc.json"
+  "browser": {
+    "console": "console-browserify",
+    "path": "path-browserify",
+    "fs": "react-native-level-fs",
+    "dgram": "react-native-udp",
+    "zlib": "browserify-zlib",
+    "constants": "constants-browserify",
+    "crypto": "react-native-crypto",
+    "dns": "dns.js",
+    "net": "react-native-tcp",
+    "domain": "domain-browser",
+    "http": "@tradle/react-native-http",
+    "https": "https-browserify",
+    "os": "react-native-os",
+    "querystring": "querystring-es3",
+    "_stream_transform": "readable-stream/transform",
+    "_stream_readable": "readable-stream/readable",
+    "_stream_writable": "readable-stream/writable",
+    "_stream_duplex": "readable-stream/duplex",
+    "_stream_passthrough": "readable-stream/passthrough",
+    "stream": "stream-browserify",
+    "timers": "timers-browserify",
+    "tty": "tty-browserify",
+    "vm": "vm-browserify",
+    "tls": false
   }
 }
