@@ -1,23 +1,11 @@
-/* eslint-disable eslint-comments/no-unlimited-disable */
-/* eslint-disable */
+/**
+ * @format
+ */
 
-// This file has been auto-generated by Ξ create-react-native-dapp Ξ.
-// Feel free to modify it, but please take care to maintain the exact
-// procedure listed between /* dapp-begin */ and /* dapp-end */, as
-// this will help persist a known template for future migrations.
-
-/* dapp-begin */
-const { Platform, LogBox } = require('react-native');
-
-if (Platform.OS !== 'web') {
-  require('react-native-get-random-values');
-  LogBox.ignoreLogs(
-    [
-      "Warning: The provided value 'ms-stream' is not a valid 'responseType'.",
-      "Warning: The provided value 'moz-chunked-arraybuffer' is not a valid 'responseType'.",
-    ],
-  );
-}
+import {AppRegistry} from 'react-native';
+import './shim.js';
+import App from './src/App';
+import {name as appName} from './app.json';
 
 if (typeof Buffer === 'undefined') {
   global.Buffer = require('buffer').Buffer;
@@ -26,21 +14,4 @@ if (typeof Buffer === 'undefined') {
 global.btoa = global.btoa || require('base-64').encode;
 global.atob = global.atob || require('base-64').decode;
 
-process.version = 'v9.40';
-
-const { registerRootComponent, scheme } = require('expo');
-const { default: App } = require('./src/App');
-
-const { default: AsyncStorage } = require('@react-native-async-storage/async-storage');
-const { withWalletConnect } = require('@walletconnect/react-native-dapp');
-
-// registerRootComponent calls AppRegistry.registerComponent('main', () => App);
-// It also ensures that whether you load the app in the Expo client or in a native build,
-// the environment is set up appropriately
-registerRootComponent(withWalletConnect(App, {
-  redirectUrl: Platform.OS === 'web' ? window.location.origin : `${scheme}://`,
-  storageOptions: {
-    asyncStorage: AsyncStorage,
-  },
-}));
-/* dapp-end */
+AppRegistry.registerComponent(appName, () => App);
