@@ -1,11 +1,11 @@
-import React, { useState } from 'react';
-import { StyleSheet, View } from 'react-native';
+import React, {useState} from 'react';
+import {StyleSheet, View} from 'react-native';
 
-import { tw } from '../ui/tailwind';
-import { Button } from '../ui/button/Button';
-import { walletInit } from '../../core/wallet';
+import {tw} from '../ui/tailwind';
+import {Button} from '../ui/button/Button';
+import {walletInit} from '../../core/wallet';
 
-export function Onboarding({ navigation }): JSX.Element {
+export function Onboarding({navigation}): JSX.Element {
   const [isLoading, setIsLoading] = useState(false);
 
   async function handleCreateWallet() {
@@ -19,8 +19,7 @@ export function Onboarding({ navigation }): JSX.Element {
       style={[
         StyleSheet.absoluteFill,
         tw`items-center justify-center bg-black px-4`,
-      ]}
-    >
+      ]}>
       <Button
         title="Create Wallet"
         onPress={handleCreateWallet}
