@@ -1,14 +1,10 @@
-import React, { useEffect, useState } from 'react';
-import { Alert, Button, StyleSheet, View } from 'react-native';
+import React, {useEffect, useState} from 'react';
+import {Alert, Button, StyleSheet, View} from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
-import { tw } from '../ui/tailwind';
+import {tw} from '../ui/tailwind';
 
-import {
-  getWalletBalance,
-  sendTransaction,
-  walletInit,
-} from '../../core/wallet';
+import {getWalletBalance, sendTransaction, walletInit} from '../../core/wallet';
 
 export function Wallet(): JSX.Element {
   async function handleCreateWallet() {
@@ -16,12 +12,12 @@ export function Wallet(): JSX.Element {
       seedPhrase: null,
       network: null,
     });
-    console.log({ newWallet });
+    console.log({newWallet});
   }
 
   async function handleShowBalance() {
     const privateKey = await AsyncStorage.getItem(
-      'WalletPrivateKey 0x6ae8018b258e978375d7e51af3d3c7891c0fcdb5'
+      'WalletPrivateKey 0x6ae8018b258e978375d7e51af3d3c7891c0fcdb5',
     );
     const walletBalance = await getWalletBalance(privateKey);
     return Alert.alert('Wallet Balance:', walletBalance);
@@ -29,10 +25,10 @@ export function Wallet(): JSX.Element {
 
   async function handleSendTransaction() {
     const privateKey = await AsyncStorage.getItem(
-      'WalletPrivateKey 0x83b7cccE2D0579ED8cA5948f082FD6cEd79DDb05'
+      'WalletPrivateKey 0x83b7cccE2D0579ED8cA5948f082FD6cEd79DDb05',
     );
-    const transactionResult = await sendTransaction({ privateKey });
-    console.log({ transactionResult });
+    const transactionResult = await sendTransaction({privateKey});
+    console.log({transactionResult});
   }
 
   async function handleDeleteAllWallets() {
@@ -40,7 +36,7 @@ export function Wallet(): JSX.Element {
     const resultAllWallets = await AsyncStorage.getAllKeys();
     console.log(
       '[resultAllWallets] - AsyncStorage allWallets',
-      resultAllWallets
+      resultAllWallets,
     );
   }
 
@@ -49,8 +45,7 @@ export function Wallet(): JSX.Element {
       style={[
         StyleSheet.absoluteFill,
         tw`flex-1 items-center justify-center bg-black`,
-      ]}
-    >
+      ]}>
       <Button title="Create wallet" onPress={handleCreateWallet} />
       <Button title="Wallet Balance" onPress={handleShowBalance} />
       <Button title="Send Transaction" onPress={handleSendTransaction} />
